Derive load command count bounds without multiplying
The count checks added with the previous bounds could be walked around by the arithmetic in the checks themselves. A segment declaring 0x40000001 sections multiplies out to 124 in 32 bits, the size of the command being checked, so the check passed and the section headers were read out of the commands after it. A build version declaring 0x20000000 tools multiplies out to nothing at all.

The largest count that fits is now derived by division, so there is no product to wrap. ComputeCommandSize computes wide and rejects a count that would not fit rather than returning a wrapped size, since it is public and a caller can reach it directly.

The same widening was missing where the reader gathers the regions the commands point at: those sizes were counts times an entry size in 32-bit arithmetic, so a wrapped product would have named a small region and left the rest of the table looking like padding.

- src/LibObjectFile/MachO/MachOSegment.cs: section count derived by division, command size computed wide
- src/LibObjectFile/MachO/MachOBuildVersionCommand.cs: tool count derived by division
- src/LibObjectFile/MachO/MachOFile.Read.cs: region sizes computed wide
- src/LibObjectFile.Tests/MachO/MachOSimpleTests.cs: the two overflow counts, which fail against the multiplying checks

+++ src/LibObjectFile/MachO/MachOFile.Read.cs
@@ -339,6 +339,9 @@ partial class MachOFile
     {
         var regions = new List<KnownRegion>();
 
+        // Sizes are counts from the file times an entry size, computed wide: a count large enough
+        // to wrap a 32-bit product would otherwise name a small region and leave the rest of the
+        // table looking like padding.
         void Add(ulong offset, ulong size, MachOSection? section = null)
         {
             // An offset of zero means the data is absent rather than at the start of the file.
@@ -364,7 +367,7 @@ partial class MachOFile
                     Add(section.FileOffset, section.Size, section);
                 }
 
-                Add(section.RelocationOffset, section.NumberOfRelocations * (ulong)MachORelocation.EntrySize);
+                Add(section.RelocationOffset, (ulong)section.NumberOfRelocations * MachORelocation.EntrySize);
             }
         }
 
@@ -377,12 +380,12 @@ partial class MachOFile
                     Add(symtab.StringOffset, symtab.StringSize);
                     break;
                 case MachODynamicSymbolTableCommand dysymtab:
-                    Add(dysymtab.TableOfContentsOffset, dysymtab.TableOfContentsCount * MachODynamicSymbolTableCommand.TableOfContentsEntrySize);
-                    Add(dysymtab.ModuleTableOffset, dysymtab.ModuleTableCount * MachODynamicSymbolTableCommand.GetModuleTableEntrySize(Is64Bit));
-                    Add(dysymtab.ExternalReferenceOffset, dysymtab.ExternalReferenceCount * MachODynamicSymbolTableCommand.ExternalReferenceEntrySize);
-                    Add(dysymtab.IndirectSymbolOffset, dysymtab.IndirectSymbolCount * MachODynamicSymbolTableCommand.IndirectSymbolEntrySize);
-                    Add(dysymtab.ExternalRelocationOffset, dysymtab.ExternalRelocationCount * MachORelocation.EntrySize);
-                    Add(dysymtab.LocalRelocationOffset, dysymtab.LocalRelocationCount * MachORelocation.EntrySize);
+                    Add(dysymtab.TableOfContentsOffset, (ulong)dysymtab.TableOfContentsCount * MachODynamicSymbolTableCommand.TableOfContentsEntrySize);
+                    Add(dysymtab.ModuleTableOffset, (ulong)dysymtab.ModuleTableCount * MachODynamicSymbolTableCommand.GetModuleTableEntrySize(Is64Bit));
+                    Add(dysymtab.ExternalReferenceOffset, (ulong)dysymtab.ExternalReferenceCount * MachODynamicSymbolTableCommand.ExternalReferenceEntrySize);
+                    Add(dysymtab.IndirectSymbolOffset, (ulong)dysymtab.IndirectSymbolCount * MachODynamicSymbolTableCommand.IndirectSymbolEntrySize);
+                    Add(dysymtab.ExternalRelocationOffset, (ulong)dysymtab.ExternalRelocationCount * MachORelocation.EntrySize);
+                    Add(dysymtab.LocalRelocationOffset, (ulong)dysymtab.LocalRelocationCount * MachORelocation.EntrySize);
                     break;
                 case MachODyldInfoCommand dyldInfo:
                     Add(dyldInfo.RebaseOffset, dyldInfo.RebaseSize);
@@ -395,7 +398,7 @@ partial class MachOFile
                     Add(data.DataOffset, data.DataSize);
                     break;
                 case MachOTwoLevelHintsCommand hints:
-                    Add(hints.Offset, hints.HintCount * MachOTwoLevelHintsCommand.HintSize);
+                    Add(hints.Offset, (ulong)hints.HintCount * MachOTwoLevelHintsCommand.HintSize);
                     break;
             }
         }
